+++ src/components/calcite-tabs/calcite-tabs.stories.ts
@@ -1,6 +1,6 @@
 import { storiesOf } from "@storybook/html";
 import { withKnobs, select } from "@storybook/addon-knobs";
-import { darkBackground, parseReadme } from "../../../.storybook/helpers";
+import { darkBackground, iconNames, parseReadme } from "../../../.storybook/helpers";
 import readme1 from "./readme.md";
 import readme2 from "../calcite-tab/readme.md";
 import readme3 from "../calcite-tab-nav/readme.md";
@@ -19,7 +19,7 @@ storiesOf("components|Tabs", module)
     "Simple",
     () => `
     <calcite-tabs
-      layout="${select("layout", { inline: "inline", center: "center" })}"
+      layout="${select("layout", ["inline", "center"])}"
     >
       <calcite-tab-nav slot="tab-nav">
         <calcite-tab-title active>Tab 1 Title</calcite-tab-title>
@@ -37,11 +37,43 @@ storiesOf("components|Tabs", module)
     { notes }
   )
   .add(
+    "With icons",
+    () => `
+    <calcite-tabs
+      layout="${select("layout", ["inline", "center"])}"
+    >
+      <calcite-tab-nav
+      slot="tab-nav">
+        <calcite-tab-title active
+        icon-start="${select("tab 1 icon-start", iconNames, iconNames[0])}"
+        >Tab 1 Title</calcite-tab-title>
+        <calcite-tab-title
+        icon-end="${select("tab 2 icon-end", iconNames, iconNames[0])}"
+        >Tab 2 Title</calcite-tab-title>
+        <calcite-tab-title
+        icon-start="${select("tab 3 icon-start", iconNames, iconNames[0])}"
+        icon-end="${select("tab 3 icon-end", iconNames, iconNames[0])}"
+        >Tab 3 Title</calcite-tab-title>
+        <calcite-tab-title
+        icon-start="${select("tab 4 icon-start", iconNames, iconNames[0])}"></calcite-tab-title>
+      </calcite-tab-nav>
+
+      <calcite-tab active><p>Tab 1 Content</p></calcite-tab>
+      <calcite-tab><p>Tab 2 Content</p></calcite-tab>
+      <calcite-tab><p>Tab 3 Content</p></calcite-tab>
+      <calcite-tab><p>Tab 4 Content</p></calcite-tab>
+    </calcite-tabs>
+  `,
+    { notes }
+  )
+  .add(
     "Dark mode",
     () => `
-    <calcite-tabs theme="dark">
+    <calcite-tabs theme="dark"
+    layout="${select("layout", ["inline", "center"])}"
+    >
       <calcite-tab-nav slot="tab-nav">
-        <calcite-tab-title active>Tab 1 Title</calcite-tab-title>
+        <calcite-tab-title active>Icon 1</calcite-tab-title>
         <calcite-tab-title>Tab 2 Title</calcite-tab-title>
         <calcite-tab-title>Tab 3 Title</calcite-tab-title>
         <calcite-tab-title>Tab 4 Title</calcite-tab-title>